CHATBOT/backend/requirements.txt
Flask
google-generativeai
PyYAML
Flask-Cors
Markdown